src/backend/requirements.txt
# Django
Django==4.2.16
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.0

# WebSocket
channels==4.0.0
channels-redis==4.2.0
daphne==4.0.0

# Database
psycopg2-binary==2.9.9

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.0
